Add the beginnings of a test for timing.c.

+++ timing.h
@@ -12,26 +12,26 @@ void timing_destroy(struct timing_struct* p_timing);
 uint64_t timing_get_total_timer_ticks(struct timing_struct* p_timing);
 uint64_t timing_get_scaled_total_timer_ticks(struct timing_struct* p_timing);
 
-size_t timing_register_timer(struct timing_struct* p_timing,
-                             void* p_callback,
-                             void* p_object);
-int64_t timing_start_timer(struct timing_struct* p_timing, size_t id);
+uint32_t timing_register_timer(struct timing_struct* p_timing,
+                               void* p_callback,
+                               void* p_object);
+int64_t timing_start_timer(struct timing_struct* p_timing, uint32_t id);
 int64_t timing_start_timer_with_value(struct timing_struct* p_timing,
-                                      size_t id,
+                                      uint32_t id,
                                       int64_t time);
-int64_t timing_stop_timer(struct timing_struct* p_timing, size_t id);
-int timing_timer_is_running(struct timing_struct* p_timing, size_t id);
-int64_t timing_get_timer_value(struct timing_struct* p_timing, size_t id);
+int64_t timing_stop_timer(struct timing_struct* p_timing, uint32_t id);
+int timing_timer_is_running(struct timing_struct* p_timing, uint32_t id);
+int64_t timing_get_timer_value(struct timing_struct* p_timing, uint32_t id);
 int64_t timing_set_timer_value(struct timing_struct* p_timing,
-                               size_t id,
+                               uint32_t id,
                                int64_t value);
 int64_t timing_adjust_timer_value(struct timing_struct* p_timing,
                                   int64_t* p_new_value,
-                                  size_t id,
+                                  uint32_t id,
                                   int64_t delta);
-int timing_get_firing(struct timing_struct* p_timing, size_t id);
+int timing_get_firing(struct timing_struct* p_timing, uint32_t id);
 int64_t timing_set_firing(struct timing_struct* p_timing,
-                          size_t id,
+                          uint32_t id,
                           int firing);
 
 int64_t timing_get_countdown(struct timing_struct* p_timing);
